Fixing modules & DB scheme

+++ system/core/models/Country.php
@@ -9,7 +9,6 @@
 
 namespace core\models;
 
-use PDO;
 use core\Model;
 use core\classes\Tool;
 use core\classes\Cache;
@@ -71,26 +70,16 @@ class Country extends Model
      */
     public function get($code)
     {
-        $country = &Cache::memory("country.$code");
-
-        if (isset($country)) {
-            return $country;
-        }
-
         $this->hook->fire('get.country.before', $code);
 
-        $sth = $this->db->prepare('SELECT * FROM country WHERE code=?');
-        $sth->execute(array($code));
-
-        $country = $sth->fetch(PDO::FETCH_ASSOC);
+        $sql = 'SELECT * FROM country WHERE code=?';
+        $options = array('unserialize' => 'format');
+        $country = $this->db->fetch($sql, array($code), $options);
 
         if (!empty($country)) {
-            $country['format'] = unserialize($country['format']);
-
             $default_format = $this->defaultFormat();
             $country['format'] = Tool::merge($default_format, $country['format']);
             $country['default'] = $this->isDefault($code);
-
             Tool::sortWeight($country['format']);
         }
 
@@ -104,78 +93,91 @@ class Country extends Model
      */
     public function defaultFormat()
     {
-        return array(
-            'country' => array(
-                'name' => $this->language->text('Country'),
-                'required' => 0, 'weight' => 0,
-                'status' => 1
-            ),
-            'state_id' => array(
-                'name' => $this->language->text('State'),
-                'required' => 0, 'weight' => 1,
-                'status' => 1
-            ),
-            'city_id' => array(
-                'name' => $this->language->text('City'),
-                'required' => 1,
-                'weight' => 2,
-                'status' => 1
-            ),
-            'address_1' => array(
-                'name' => $this->language->text('Address'),
-                'required' => 1,
-                'weight' => 3,
-                'status' => 1
-            ),
-            'address_2' => array(
-                'name' => $this->language->text('Additional address'),
-                'required' => 0,
-                'weight' => 4,
-                'status' => 0
-            ),
-            'phone' => array(
-                'name' => $this->language->text('Phone'),
-                'required' => 1,
-                'weight' => 5,
-                'status' => 1
-            ),
-            'postcode' => array(
-                'name' => $this->language->text('Post code'),
-                'required' => 1,
-                'weight' => 6,
-                'status' => 1
-            ),
-            'first_name' => array(
-                'name' => $this->language->text('First name'),
-                'required' => 1,
-                'weight' => 7,
-                'status' => 1
-            ),
-            'middle_name' => array(
-                'name' => $this->language->text('Middle name'),
-                'required' => 1,
-                'weight' => 8,
-                'status' => 1
-            ),
-            'last_name' => array(
-                'name' => $this->language->text('Last name'),
-                'required' => 1,
-                'weight' => 9,
-                'status' => 1
-            ),
-            'company' => array(
-                'name' => $this->language->text('Company'),
-                'required' => 0,
-                'weight' => 10,
-                'status' => 0
-            ),
-            'fax' => array(
-                'name' => $this->language->text('Fax'),
-                'required' => 0,
-                'weight' => 11,
-                'status' => 0
-            ),
+        $items = array();
+
+        $items['country'] = array(
+            'name' => $this->language->text('Country'),
+            'required' => 0, 'weight' => 0,
+            'status' => 1
         );
+
+        $items['state_id'] = array(
+            'name' => $this->language->text('State'),
+            'required' => 0, 'weight' => 1,
+            'status' => 1
+        );
+
+        $items['city_id'] = array(
+            'name' => $this->language->text('City'),
+            'required' => 1,
+            'weight' => 2,
+            'status' => 1
+        );
+
+        $items['address_1'] = array(
+            'name' => $this->language->text('Address'),
+            'required' => 1,
+            'weight' => 3,
+            'status' => 1
+        );
+
+        $items['address_2'] = array(
+            'name' => $this->language->text('Additional address'),
+            'required' => 0,
+            'weight' => 4,
+            'status' => 0
+        );
+
+        $items['phone'] = array(
+            'name' => $this->language->text('Phone'),
+            'required' => 1,
+            'weight' => 5,
+            'status' => 1
+        );
+
+        $items['postcode'] = array(
+            'name' => $this->language->text('Post code'),
+            'required' => 1,
+            'weight' => 6,
+            'status' => 1
+        );
+
+        $items['first_name'] = array(
+            'name' => $this->language->text('First name'),
+            'required' => 1,
+            'weight' => 7,
+            'status' => 1
+        );
+
+        $items['middle_name'] = array(
+            'name' => $this->language->text('Middle name'),
+            'required' => 1,
+            'weight' => 8,
+            'status' => 1
+        );
+
+        $items['last_name'] = array(
+            'name' => $this->language->text('Last name'),
+            'required' => 1,
+            'weight' => 9,
+            'status' => 1
+        );
+
+        $items['company'] = array(
+            'name' => $this->language->text('Company'),
+            'required' => 0,
+            'weight' => 10,
+            'status' => 0
+        );
+
+        $items['fax'] = array(
+            'name' => $this->language->text('Fax'),
+            'required' => 0,
+            'weight' => 11,
+            'status' => 0
+        );
+
+        return $items;
     }
 
     /**
@@ -241,6 +243,7 @@ class Country extends Model
         $result = true;
         $this->db->insert('country', $data);
         $this->hook->fire('add.country.after', $data, $result);
+
         return (bool) $result;
     }
 
@@ -289,28 +292,28 @@ class Country extends Model
             return false;
         }
 
-        $this->db->delete('country', array('code' => $code));
-        $this->db->delete('zone', array('country' => $code));
-        $this->db->delete('city', array('country' => $code));
-        $this->db->delete('state', array('country' => $code));
+        $result = (bool) $this->db->delete('country', array('code' => $code));
 
-        $this->hook->fire('delete.country.after', $code);
-        return true;
+        if ($result) {
+            $this->db->delete('zone', array('country' => $code));
+            $this->db->delete('city', array('country' => $code));
+            $this->db->delete('state', array('country' => $code));
+        }
+
+        $this->hook->fire('delete.country.after', $code, $result);
+        return (bool) $result;
     }
 
     /**
-     * Returns true if a country can be deleted
+     * Returns true if the country can be deleted
      * @param string $code
      * @return boolean
      */
     public function canDelete($code)
     {
         $sql = 'SELECT address_id FROM address WHERE country=?';
+        $result = $this->db->fetchColumn($sql, array($code));
 
-        $sth = $this->db->prepare($sql);
-        $sth->execute(array($code));
-
-        $result = $sth->fetchColumn();
         return empty($result);
     }
 
@@ -388,14 +391,11 @@ class Country extends Model
             $sql .= ' LIMIT ' . implode(',', array_map('intval', $data['limit']));
         }
 
-        $sth = $this->db->prepare($sql);
-        $sth->execute($where);
-
         if (!empty($data['count'])) {
-            return (int) $sth->fetchColumn();
+            return (int) $this->db->fetchColumn($sql, $where);
         }
 
-        $results = $sth->fetchAll(PDO::FETCH_ASSOC);
+        $results = $this->db->fetchAll($sql, $where);
 
         $list = array();
         foreach ($results as $country) {
